Preparazione Esame

+++ Classroom/iTunes/model/album.py
@@ -5,10 +5,13 @@ class Album:
     AlbumId: int
     Title: str
     ArtistId: int
-    totD: int
+    totD: float
 
     def __hash__(self):
         return hash(self.AlbumId)
+
+    #def __eq__(self, other):
+        #return self.AlbumId == other.AlbumId
 
     def __str__(self):
         return f"{self.Title} -- {toMinutes(self.totD)}"
